+++ crates/flowctl/src/preview/driver.rs
@@ -1,19 +1,10 @@
-//! Capture driver for `flowctl preview-next`.
-//!
-//! Captures are leaderless: each shard runs its own connector container,
-//! RocksDB, publish loop, and transaction loop with no cross-shard coordination.
-//! The preview driver mirrors that — N synthetic shards each driving one long-lived
-//! SessionLoop via `runtime_next::shard::Service::spawn_capture`, fanned out as
-//! independent `tokio::spawn` tasks.
-//!
-//! Real-world note: many connectors expect N=1 today, and a connector that
-//! ignores `Open.range.key_begin/key_end` will duplicate work across shards.
-//! Same caveat as the materialize driver — N>1 exercises the runtime, but
-//! interpreting the output as a workload signal requires a range-partitioning
-//! connector.
+//! Run-scoped driver: spawns N shard tasks via
+//! `runtime_next::shard::Service::spawn_materialize`, synthesizing the
+//! SessionLoop/Join/Task envelopes the controller (Go in production) would
+//! normally send.
 
-use crate::raw::preview_next::Controls;
-use crate::raw::preview_next::services::Run;
+use crate::preview::Controls;
+use crate::preview::services::Run;
 use anyhow::Context;
 use prost::Message;
 use proto_flow::{flow, runtime as cruntime};
@@ -22,15 +13,18 @@ use tokio::sync::mpsc;
 use tokio_stream::wrappers::UnboundedReceiverStream;
 use tokio_util::sync::CancellationToken;
 
+/// Run preview sessions against the prepared topology. Sessions are
+/// numbered `1..` for log context, and all run over the same per-shard
+/// SessionLoop streams.
 pub async fn run_sessions(
     run: &Run,
-    spec: &flow::CaptureSpec,
+    spec: &flow::MaterializationSpec,
     session_targets: Vec<u32>,
+    fixture_dirs: Vec<String>,
     controls: Controls,
     stop_token: CancellationToken,
 ) -> anyhow::Result<()> {
-    let join_shards =
-        crate::raw::preview_next::shards::build_capture_join_shards(run.n_shards, spec)?;
+    let join_shards = crate::preview::shards::build_materialize_join_shards(run.n_shards, spec)?;
     // Encode the spec once; each shard's Task carries a cheap refcount clone of
     // these bytes rather than deep-cloning and re-encoding the spec per shard.
     let spec_bytes: bytes::Bytes = spec.encode_to_vec().into();
@@ -38,16 +32,16 @@ pub async fn run_sessions(
     let mut handles = Vec::with_capacity(run.n_shards as usize);
     for i in 0..run.n_shards {
         let run_handle = RunHandle {
-            // Shard 0 uses the Run's tracked tempdir so it's surfaced in the
-            // startup log and observable post-run. Shards >=1 each get their
-            // own auto-managed tempdir via RocksDB::open(None).
-            rocksdb_path: (i == 0).then(|| run.rocksdb_path.clone()),
+            peer_endpoint: run.peer_endpoint.clone(),
+            shuffle_log_dir: run.shuffle_log_dir.clone(),
+            rocksdb_path: run.rocksdb_path.clone(),
             network: run.network.clone(),
             registry: run.registry.clone(),
         };
         let spec_bytes = spec_bytes.clone();
-        let join_shard = join_shards[i as usize].clone();
+        let join_shards = join_shards.clone();
         let session_targets = session_targets.clone();
+        let fixture_dirs = fixture_dirs.clone();
         let controls = controls.clone();
         let stop_token = stop_token.clone();
 
@@ -56,8 +50,9 @@ pub async fn run_sessions(
                 run_handle,
                 spec_bytes,
                 i,
-                join_shard,
+                join_shards,
                 session_targets,
+                fixture_dirs,
                 controls,
                 stop_token,
             )
@@ -65,19 +60,20 @@ pub async fn run_sessions(
         }));
     }
 
-    // First error wins; remaining shards observe their request channel dropping
-    // on handle drop and tear down naturally.
+    // Await all shard drivers. The first error surfaces; remaining drivers
+    // observe their request channel dropping (when their handle drops) and
+    // tear down naturally.
     let mut first_err: Option<anyhow::Error> = None;
     for handle in handles {
         match handle.await {
             Ok(Ok(())) => (),
             Ok(Err(e)) if first_err.is_none() => first_err = Some(e),
-            Ok(Err(e)) => tracing::warn!(error = ?e, "secondary capture shard driver error"),
+            Ok(Err(e)) => tracing::warn!(error = ?e, "secondary shard driver error"),
             Err(panic) => {
                 if first_err.is_none() {
-                    first_err = Some(anyhow::anyhow!("capture driver panic: {panic}"));
+                    first_err = Some(anyhow::anyhow!("driver panic: {panic}"));
                 } else {
-                    tracing::warn!(?panic, "secondary capture driver panic");
+                    tracing::warn!(?panic, "secondary driver panic");
                 }
             }
         }
@@ -88,10 +84,12 @@ pub async fn run_sessions(
     Ok(())
 }
 
-/// `Run` fields a single capture shard driver needs. Cheaper to clone than
-/// `&Run` so we can move it into a spawned task without lifetime gymnastics.
+/// `Run` fields a single shard driver needs. Cheaper to clone than `&Run`
+/// so we can hand it into a spawned task without lifetime gymnastics.
 struct RunHandle {
-    rocksdb_path: Option<String>,
+    peer_endpoint: String,
+    shuffle_log_dir: String,
+    rocksdb_path: String,
     network: String,
     registry: service_kit::Registry,
 }
@@ -100,16 +98,19 @@ async fn drive_one_shard(
     run: RunHandle,
     spec_bytes: bytes::Bytes,
     shard_index: u32,
-    join_shard: proto::join::Shard,
+    join_shards: Vec<proto::join::Shard>,
     session_targets: Vec<u32>,
+    fixture_dirs: Vec<String>,
     controls: Controls,
     stop_token: CancellationToken,
 ) -> anyhow::Result<()> {
-    let task_name = format!("preview-capture-{shard_index:03}");
+    let (request_tx, request_rx) = mpsc::unbounded_channel::<tonic::Result<proto::Materialize>>();
+
+    let task_name = format!("preview-shard-{shard_index:03}");
 
     let shard_svc = runtime_next::shard::Service::new(
         cruntime::Plane::Local,
-        run.network,
+        run.network.clone(),
         None,
         task_name,
         controls.publisher_factory.clone(),
@@ -118,32 +119,34 @@ async fn drive_one_shard(
         None, // No AuthN+AuthZ signer (local loopback).
     );
 
-    let (request_tx, request_rx) = mpsc::unbounded_channel::<tonic::Result<proto::Capture>>();
-    let mut response_rx = shard_svc.spawn_capture(UnboundedReceiverStream::new(request_rx));
+    let mut response_rx = shard_svc.spawn_materialize(UnboundedReceiverStream::new(request_rx));
 
     // Seed shard zero's RocksDB with any `--initial-state` before the runtime
     // opens it at SessionLoop, so it recovers the state on its first scan.
-    // Only shard zero carries a tracked `rocksdb_path`.
-    if let Some(rocksdb_path) = &run.rocksdb_path {
-        if !controls.initial_state_json.is_empty() {
-            super::seed_preview_state(
-                cruntime::RocksDbDescriptor {
-                    rocksdb_path: rocksdb_path.clone(),
-                    rocksdb_env_memptr: 0,
-                },
-                &controls.initial_state_json,
-            )
-            .await
-            .context("seeding --initial-state into shard-zero RocksDB")?;
-        }
+    if shard_index == 0 && !controls.initial_state_json.is_empty() {
+        super::seed_preview_state(
+            cruntime::RocksDbDescriptor {
+                rocksdb_path: run.rocksdb_path.clone(),
+                rocksdb_env_memptr: 0,
+            },
+            &controls.initial_state_json,
+        )
+        .await
+        .context("seeding --initial-state into shard-zero RocksDB")?;
     }
 
-    let rocksdb_descriptor = run.rocksdb_path.map(|p| cruntime::RocksDbDescriptor {
-        rocksdb_path: p,
-        rocksdb_env_memptr: 0,
-    });
+    // Open the SessionLoop once. `runtime-next` opens RocksDB here and keeps
+    // the handle live across the repeated Join/Task sessions below.
+    let rocksdb_descriptor = if shard_index == 0 {
+        Some(cruntime::RocksDbDescriptor {
+            rocksdb_path: run.rocksdb_path.clone(),
+            rocksdb_env_memptr: 0,
+        })
+    } else {
+        None
+    };
     request_tx
-        .send(Ok(proto::Capture {
+        .send(Ok(proto::Materialize {
             session_loop: Some(proto::SessionLoop { rocksdb_descriptor }),
             ..Default::default()
         }))
@@ -155,48 +158,48 @@ async fn drive_one_shard(
         }
         let session_index = idx + 1;
 
+        // A fixture preview reads each session from its own directory (fresh
+        // segments from segment one); live preview shares the run's directory.
+        let shuffle_directory = fixture_dirs
+            .get(idx)
+            .cloned()
+            .unwrap_or_else(|| run.shuffle_log_dir.clone());
+
         request_tx
-            .send(Ok(proto::Capture {
+            .send(Ok(proto::Materialize {
                 join: Some(proto::Join {
                     etcd_mod_revision: session_index as i64,
-                    shards: vec![join_shard.clone()],
-                    shard_index: 0,
-                    // Captures have no shuffle or leader; the handler ignores these.
-                    shuffle_directory: String::new(),
-                    shuffle_endpoint: String::new(),
-                    leader_endpoint: String::new(),
+                    shards: join_shards.clone(),
+                    shard_index,
+                    shuffle_directory,
+                    shuffle_endpoint: run.peer_endpoint.clone(),
+                    leader_endpoint: run.peer_endpoint.clone(),
                 }),
                 ..Default::default()
             }))
             .map_err(|_| anyhow::anyhow!("serve task closed before Join"))?;
 
+        // All shards receive Task. Shard zero alone forwards to the leader.
         tracing::info!(
-            shard_index,
             session = session_index,
+            shard_index,
             target_txns,
-            "starting preview capture session",
+            "starting preview session",
         );
 
         request_tx
-            .send(Ok(proto::Capture {
+            .send(Ok(proto::Materialize {
                 task: Some(proto::Task {
                     spec: spec_bytes.clone(),
                     max_transactions: target_txns,
                     sqlite_vfs_uri: String::new(),
-                    publisher_id: Default::default(),
+                    publisher_id: Default::default(), // The harness forwards no leader producer.
                 }),
                 ..Default::default()
             }))
             .map_err(|_| anyhow::anyhow!("serve task closed before Task"))?;
 
-        drive_session_responses(
-            &request_tx,
-            &mut response_rx,
-            shard_index,
-            session_index,
-            &stop_token,
-        )
-        .await?;
+        drive_session_responses(&request_tx, &mut response_rx, session_index, &stop_token).await?;
     }
 
     drop(request_tx);
@@ -208,13 +211,12 @@ async fn drive_one_shard(
 }
 
 async fn drive_session_responses(
-    request_tx: &mpsc::UnboundedSender<tonic::Result<proto::Capture>>,
-    response_rx: &mut mpsc::UnboundedReceiver<tonic::Result<proto::Capture>>,
-    shard_index: u32,
+    request_tx: &mpsc::UnboundedSender<tonic::Result<proto::Materialize>>,
+    response_rx: &mut mpsc::UnboundedReceiver<tonic::Result<proto::Materialize>>,
     session_index: usize,
     stop_token: &CancellationToken,
 ) -> anyhow::Result<()> {
-    let verify = runtime_next::verify("Capture", "Joined, Opened, or Stopped", "shard");
+    let verify = runtime_next::verify("Materialize", "Joined, Opened, or Stopped", "shard");
 
     let mut requested_stop = false;
     loop {
@@ -224,7 +226,7 @@ async fn drive_session_responses(
             _ = stop_token.cancelled(), if !requested_stop => {
                 requested_stop = true;
                 _ = request_tx
-                    .send(Ok(proto::Capture {
+                    .send(Ok(proto::Materialize {
                         stop: Some(proto::Stop {}),
                         ..Default::default()
                     }));
@@ -233,16 +235,11 @@ async fn drive_session_responses(
                 let msg = verify.not_eof(msg)?;
 
                 if let Some(proto::Joined { max_etcd_revision }) = msg.joined {
-                    if max_etcd_revision != 0 {
-                        anyhow::bail!(
-                            "capture preview expected Joined.max_etcd_revision = 0, got {max_etcd_revision}",
-                        );
-                    }
-                    tracing::debug!(shard_index, session_index, "capture session joined");
-                } else if let Some(proto::capture::Opened { container }) = &msg.opened {
-                    tracing::debug!(shard_index, session_index, ?container, "capture session opened");
+                    tracing::debug!(session_index, max_etcd_revision, "session joined");
+                } else if let Some(proto::materialize::Opened { container, .. }) = &msg.opened {
+                    tracing::debug!(session_index, ?container, "session opened");
                 } else if let Some(proto::Stopped {}) = msg.stopped {
-                    tracing::debug!(shard_index, session_index, "capture session stopped");
+                    tracing::debug!(session_index, "session stopped");
                     return Ok(());
                 } else {
                     return Err(verify.fail_msg(msg));
