+++ crates/runtime/src/lib.rs
@@ -6,7 +6,6 @@ mod capture;
 mod combine;
 mod container;
 mod derive;
-pub mod harness;
 mod image_connector;
 mod local_connector;
 mod materialize;
@@ -308,47 +307,6 @@ impl Verify {
             ))
         }
     }
-}
-
-/// exchange is a combinator for avoiding deadlocks. It sends into a request
-/// Stream while concurrently polling and yielding responses of a corresponding
-/// response Stream. It returns a stream which completes once the send has
-/// completed.
-///
-/// `exchange` mitigates an extremely common deadlock mistake, of sending into
-/// a receiver without consideration for whether the receiver may be unable to
-/// receive because it's output channel is stuffed and is not being serviced.
-/// This is a generalized problem -- in no way unique to Rust -- but the polled
-/// nature of Futures and Streams accentuate it because receiving from a Stream
-/// is *also* polling it, allowing it to perform other important activity even
-/// if it cannot immediately yield an item.
-fn exchange<'s, Request, Tx, Response, Rx>(
-    request: Request,
-    tx: &'s mut Tx,
-    rx: &'s mut Rx,
-) -> impl futures::Stream<Item = Response> + 's
-where
-    Request: 'static,
-    Tx: futures::Sink<Request> + Unpin + 's,
-    Rx: futures::Stream<Item = Response> + Unpin + 's,
-{
-    use futures::{SinkExt, StreamExt};
-
-    futures::stream::unfold((tx.feed(request), rx), move |(mut feed, rx)| async move {
-        tokio::select! {
-            biased;
-
-            // We suppress a `feed` error, which represents a disconnection / reset,
-            // because a better and causal error will invariably be surfaced by `rx`.
-            _result = &mut feed => None,
-
-            response = rx.next() => if let Some(response) = response {
-                Some((response, (feed, rx)))
-            } else {
-                None
-            },
-        }
-    })
 }
 
 fn truncate_chars(s: &str, max_chars: usize) -> &str {
